MATTER-2644 Fix ble-wifi pairing timeout issue
By setting the AdvertisementInterval of BLE to 512,
the interval can be set as 0.625ms * 512 = 320ms.
When the adv is enabled on the cluster end,
its advertisement interval is 320s instead of the default value 1280ms.
This can make the BLE connection take less time.

Integrate the main.conf and install it to path /etc/bluetooth/.

Change-Id: Icc625e32de95bbd56f33c4217befca7b46064483
Signed-off-by: LiJunmeng <junmeng.li_1@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/25708
Reviewed-by: Elven Wang <elven.wang@nxp.com>